+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add `SessionManager` and `ModelManager` classes
 - discourage the use of Active Record anti-patterns (remove `query` attribute from `BaseModel`)
 - bugfix for `declarative_base` if a custom base model is passed in without a constructor
+- configure the `MetaData` naming convention if none is provided
 
 ## v0.5.1 (2018/10/20)